EKM/requirements.txt
asn1crypto==0.23.0
bcrypt==3.1.4
beefish==0.1.2
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.1
cryptography==2.1.2
Django==1.11.6
django-after-response==0.2.2
django-cors-headers==2.1.0
django-filter==1.1.0
django-rest-swagger==2.1.2
djangorestframework==3.7.1
enum34==1.1.6
Fabric==1.14.0
idna==2.6
ipaddress==1.0.18
itypes==1.1.0
jenkinsapi==0.3.4
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
openapi-codec==1.3.2
paramiko==2.3.1
psycopg2==2.7.3.1
pyasn1==0.3.7
pycparser==2.18
#pycrypto==2.6.1
PyNaCl==1.1.2
pytz==2017.2
redis==2.10.6
requests==2.18.4
serpy==0.2.0
simplejson==3.11.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
xmljson==0.1.9
